Move static methods to new Convert class to avoid circular imports and make them more widely available and testable.
